Added refactor
